Wired up the info from the currently-centered video to pipe it into the <Info> widget. Ref #21
Introduced a class wrapper for the data entries in layout.json to add
behaviors to the data such as extracting data for <Info>

This commit also has a bunch of JS formatting changes (e.g. ensure
semicolons at the ends of lines) because apparently our text editor
auto-save formatters are still disagreeing slightly. ¯\_(ツ)_/¯