Stop allowing the user to pinch-zoom beyond the minimum or maximum zoom with elastic snap-back, so that we don't accidentally load a ton of videos and crash when they zoom way out.
Adjust the zoom levels so that users can zoom in further on mobile to
compensate for taking away the elastic snap-back.

Add a 1px border around every video that is 100% transparent. No idea
why, but this seems to mitigate the crackle of black borders around the
videos during mobile zooming.

Offset the initial position of the first tardigrade video slightly so
that it doesn't clash with the text on an iPhone X size screen.